1) Added unit tests for helper functions \n 2) Moved helper funtion utilities to separate file
