+++ tox-integration.ini
@@ -22,7 +22,6 @@ deps =
     paho-mqtt
     mqtt: fluent-logger
     pytest-remove-stale-bytecode
-    jmespath
 commands =
     docker-compose stop
     docker-compose build
